Fix audit findings: input bounds, fatal flags, solver guards, warnings, unit tests
  - Bound history input reading (no 1000-row UB), fatal on empty history
  - Make out-of-range input_settings flags fatal (Switch -> log + stderr + exit)
  - Fix QuarticEquation convergence test (fabs), warn on Newton non-convergence
  - Guard grain-radius division; fatal on invalid gas in getDiffusionModes*
  - Meyers singleton for Simulation; named size/physics constants
  - Enable -Wall -Wextra (tree warning-clean), fix PROJECTSOUR typo, C++17 explicit
  - Add unit tests (Solver, SciantixArray) wired into CTest

  Regression: 109/109 PASS, gold unchanged. TU-coupling build verified.

+++ include/classes/SciantixArray.h
@@ -75,7 +75,7 @@ template <class T> class SciantixArray
     {
         array = data;
 
-        for (int i = 0; i < data.size(); i++)
+        for (size_t i = 0; i < data.size(); i++)
         {
             map[data[i].getName()] = i;
         }
